viewport fix and touchups

+++ src/App.js
@@ -1,27 +1,10 @@
 import React, { Component, useState } from "react";
 import data from "./shrek.json";
-import { css } from 'linaria';
 import { Router } from "@reach/router";
 import Menu from "./Menu";
 import Video from "./Video";
 import Settings from "./Settings";
-
-const marqueeStyles = css`
-  position: fixed;
-  width: 100vw;
-  bottom: 0px;
-  left: 0px;
-  padding: 2px 4px;
-  background-color: #999;
-  img {
-    margin-top: -2px;
-    margin-bottom: -2px;
-    margin-left: 4px;
-    height: 24px;
-    filter: brightness(0);
-    box-shadow: inset 0px -2px 0px 0px black;
-  }
-`;
+import Info from "./Info";
 
 const buildPageComponent = data => (props) => {
   const {
@@ -64,16 +47,48 @@ const debugConfig = {
   sourceDir: '/static/video/',
 };
 
+const getResizeValues = ({ width, height }) => {
+  const rect = document.getElementById('app').getBoundingClientRect();
+  const scaleByWidth = rect.width / width;
+  const scaleByHeight = rect.height / height;
+  const scale = scaleByWidth > scaleByHeight ? scaleByHeight : scaleByWidth;
+  return { transform: `scale(${scale})`, transformOrigin: 'center left' };
+}
+
 class App extends Component {
   constructor(props) {
     super(props);
     this.ohDearItsIos = /iPhone|iPod|iPad/.test(navigator.platform);
+    this.state = {
+      style: document.getElementById('app').getBoundingClientRect().width < 800 ? getResizeValues({ width: 853, height: 480 }) : undefined
+    };
+  }
+
+  ref;
+  resizeListener = () => {
+    const style = getResizeValues({ width: 853, height: 480 });
+    if (style.transform) {
+      return;
+    }
+    const change = Math.abs(+style.transform.split(/\(|\)/)[1] - this.state.style.transform.split(/\(|\)/)[1]);
+    if (change > 0.2) {
+      this.setState({ style })
+    }
   }
 
   componentDidMount() {
     let iOSNote = 'IOS_WARNING';
-    if (this.ohDearItsIos && !window.localStorage.getItem(iOSNote) && confirm('Warning: There will be video rendering issues on iOS')) {
+    if (this.ohDearItsIos && !window.localStorage.getItem(iOSNote) && confirm('Warning: There will be video rendering issues on iOS. Please refresh if video fails and double click buttons.')) {
       window.localStorage.setItem(iOSNote, true);
+    }
+    if (this.state.style) {
+      window.addEventListener('resize', this.resizeListener);
+    }
+  }
+
+  componentWillUnmount() {
+    if (this.listener) {
+      window.removeEventListener('resize', this.resizeListener)
     }
   }
 
@@ -82,40 +97,16 @@ class App extends Component {
     const sceneData = buildScenesPages(scenes);
     const pageOptions = { ...pages, ...sceneData };
     let links = [];
-    const footer = (
-      <div className={marqueeStyles}>
-        Hi, this is a recreation of a DVD menu using the source menu file from the DVD. {' '}
-        Most custom configuration is handled via JSON so the code could be used to make others. {' '}<br />
-        Apologies for the video quality, I'm not very good at converting dvd menus to mp4
-        <br />
-        <p>
-          Other things I've hacked together include:{' '}
-          {[
-            { link: 'https://packard-belle.netlify.com', text: 'Windows98 Clone', github: 'https://github.com/padraigfl/packard-belle' },
-            { link: 'https://react-coursebuilder.netlify.com', text: 'Youtube Playlists\' Note Taker', github: 'https://github.com/padraigfl/videojs-react-course-assistant' },
-            { link: 'https://omdb--critics-lists.netlify.com/', text: 'End-of-year critics\' lists breakdown', github: 'https://github.com/padraigfl/critics-lists' },
-            { link: 'https://github.com/padraigfl/us-bus-data', text: 'GeoJSON data for loads of US public transit routes' }
-          ].map((v, idx, arr) => (
-            <>
-              <a href={v.link} target="_blank">{v.text}</a>
-              {v.github && (
-                <a href={v.github} target="_blank"><img src="/static/highlights/github-icon.png" /></a>
-              )}
-              {idx !== arr.length - 1 && `, `}
-            </>
-          ))}
-        </p>
-        <p>
-          If you'd like to improve the videojs integration or add some other menus that'd be really neat!{' '}
-          [<a href="https://github.com/padraigfl/dvd-menu" target="_blank">Source Code</a>]
-        </p>
-      </div>
-    );
-
     return (
       <>
-        {footer}
-        <div id="wrapper" className={config.title} style={ window.innerWidth < 500 ? { transform: `scale(${window.innerWidth / 853})`, transformOrigin: 'center left' } : undefined}>
+        <a
+          href="https://github.com/padraigfl/dvd-menu"
+          target="_blank"
+          style={ {position: 'absolute', top: '0', left: '0' } }
+        >
+          <img style={{width: '32px', height: '32px'}} src="/static/highlights/github-icon.png"/>
+        </a>
+        <div id="wrapper" ref={this.ref} className={config.title} style={this.state.style}>
           <Video
             launch={launch}
             config={{
@@ -137,6 +128,7 @@ class App extends Component {
                           startPageChange={startPageChange}
                           path={link}
                           key={`r${idx}`}
+                          default={link === 'launch'}
                         />
                       );
                     })}
@@ -149,6 +141,7 @@ class App extends Component {
           <Link to={link}> {link}</Link>
         ))} */}
         </div>
+        <Info />
       </>
     );
   }
